Create Post service
